LayoutTests/webcl/tests/webcl-test-context-properties-expected.txt

Testing WebCLContextProperties

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS devices.length > 0 is true
PASS devices[0] instanceof WebCLDevice is true
PASS platform instanceof WebCLPlatform is true
PASS deviceType === cl.DEVICE_TYPE_DEFAULT is true
PASS successfullyParsed is true

TEST COMPLETE
